+++ data-raw/parameters_starting_values.R
@@ -34,7 +34,7 @@ default_parameters <- list(
 
   # fishpop movement
   move_mean = 10.0,
-  move_var = 5.0,
+  move_sd = 5.0,
   move_reef = 1.0,
   move_return = 15.0,
   move_border = 1.0,
@@ -49,7 +49,7 @@ default_parameters <- list(
   # fishpop reserves
   pop_reserves_max = 0.05,
   pop_reserves_thres_mean = 0.1,
-  pop_reserves_thres_var = 0.0,
+  pop_reserves_thres_sd = 0.0,
   pop_reserves_consump = 0.1,
 
   # fishpop respiration
@@ -79,7 +79,7 @@ default_starting <- list(
   # fishpop related
   pop_n = 8,
   pop_mean_size = 9.0,
-  pop_var_size = 10.0
+  pop_sd_size = 10.0
 )
 
 usethis::use_data(default_parameters, default_starting, overwrite = TRUE)