+++ src/rcpp_move_wrap.cpp
@@ -16,7 +16,7 @@ using namespace Rcpp;
 //' @param fishpop Matrix with fishpop values.
 //' @param fishpop_attr Matrix with id and threshold of pop_reserves_max.
 //' @param movement String specifing movement algorithm.
-//' @param move_mean,move_var Double with mean movement parameter.
+//' @param move_mean,move_sd Double with mean movement parameter.
 //' @param move_reef Double with mean movement distance when sheltering at reef.
 //' @param move_border Double with movement distance that surrounds reef cell border.
 //' @param move_return Double with mean movement distance when returning to reef.
@@ -39,7 +39,7 @@ using namespace Rcpp;
 //' @keywords internal
 // [[Rcpp::export]]
 void rcpp_move_wrap(Rcpp::NumericMatrix fishpop, Rcpp::NumericMatrix fishpop_attr,
-                    std::string movement, double move_mean, double move_var, double move_reef,
+                    std::string movement, double move_mean, double move_sd, double move_reef,
                     double move_border, double move_return, double max_dist,
                     Rcpp::NumericMatrix coords_reef, Rcpp::NumericVector extent,
                     Rcpp::IntegerVector dimensions) {
@@ -47,19 +47,19 @@ void rcpp_move_wrap(Rcpp::NumericMatrix fishpop, Rcpp::NumericMatrix fishpop_att
   // random movement
   if (movement == "rand") {
 
-    rcpp_move_rand(fishpop, move_mean, move_var, max_dist, FALSE,
+    rcpp_move_rand(fishpop, move_mean, move_sd, max_dist, FALSE,
                    coords_reef, extent, dimensions);
 
     // attracted movement
   } else if (movement == "attr") {
 
-    rcpp_move_rand(fishpop, move_mean, move_var, max_dist, true,
+    rcpp_move_rand(fishpop, move_mean, move_sd, max_dist, true,
                    coords_reef, extent, dimensions);
 
     // behaviour movement
   } else if (movement == "behav") {
 
-    rcpp_move_behav(fishpop, fishpop_attr, move_mean, move_var,
+    rcpp_move_behav(fishpop, fishpop_attr, move_mean, move_sd,
                     move_reef, move_border, move_return, max_dist,
                     coords_reef, extent, dimensions);