+++ NEWS.md
@@ -1,6 +1,7 @@
 # arrR 1.4.3
 * Allow to write results into a text file
 * Renamed `get_stable_values` to `get_req_nutrients` and changes to function
+* Renamed all `*_var` parameters to `*_sd`
 
 # arrR 1.4.2.1
 * Better structure of `rcpp_fishpop_growth()`